requirements.txt
torch
torchvision
torchaudio
numpy>=1.24.3
Pillow>=10.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
tqdm>=4.66.1